requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
wordcloud
kaggle
sentence_transformers
psycopg2-binary
python-dotenv
SQLAlchemy
psycopg2